Implemented successful Waist strategy
First iteration of successful waist strategy has been implemented. It
allows the legs to balance the hips in a central position while
standing. Currently all joints bar waist are locked so there is no leg
wobble to test against. However this has proved to not be 100% accurate
as with a stiff incline the legs fall backwards into a sitting pose
keeping the torso balanced forwards while falling. To compensate I will
need to have extensions that allow the waist to lean past its natural
balance point to ensure it does not topple. As such I may need a set of
two values and inverse kinetic angle to understand falling when using
the absolute orientation to balance.