Completed Waist strategy, started Abs strategy
The legs can now balance at an incline of up to 0.45 unassisted. Next
balance for the abs will be added. After that I will either start
looking at the Kinect for controlling what I have or will look to add
centre of momentum and start work with the logic behind that.