Finished Kinect controllers now need motor controllers
Resolved a few issues with the robot however jitter from an undefined
source is causing some issues.
Now that the kinect controller is working nicely we can look at either
slerping the spheres to the controllers or starting on motor updating
and limits as these have been reset to static for some time now.