Leg's motion 1
Been a while since an update however a lot of progress has been made!
Now I have a set of legs that accurately move with my own. That is bar
lifting my thigh. There is an issue with there being no other accurate
reference for these joints in comparison to the rest of the body.
I have a joint for the hip, and knee, however I need one more which
would be relative to the hip on a flat plane.

Howevere issues with the kinects inability to understand rotation could
cause some trouble. I may be able to resolve this by making the spine a
fixed point and calculating the spheres rotations around this.